Create Number Guessing Game
This task is about guessing a random number by the user. It pops up Congratulations! The Number you guessed is Right! when one guesses it right. It also shows whether the guessed number is less than or greater than random number. Most importantly it should be guessed in given number of trials only.